Add golang kafka producer/consumer
It uses Schema Registry to verify if the schema is registered. It won't
register a new schema.

A custom serde is required in order to skip the magical byte and fetch
the schema id from each message.

This approach utilizes a specific avro schema. This works well if you
have base schema that you need to always work with. In this way you can
work with a struct instead of a generic map.

There are other ways of consuming that don't require a specific (struct)
for each event.